Remove string as valid argument for insertOne
Writer::insertOne accepts a string as a valid argument.
When a string was encountered the method tried to parse it as a CSV. This
parsing would result in unexpected data if one was injecting wrong data.

By removing string conversion we improve method expectation as well as
the method performance.