Reader and ResultSet record offset preservation
The Reader and the ResultSet object must keep as much as possible
information from the original CSV document. To do so we are:

- Removing ResultSet::preserveRecordOffset
- Removing ResultSet::isRecordOffsetPreserved
- Making sure that ResultSet::fetchColumn and ResultSet::getRecords
  always return the CSV document original record offset.

Because this is expected, records offset will not be preserved on JSON conversion.
To preserved them you are required to directly use getRecords with iterator_to_array

- renamed internal Document to Stream
- simplified Exception handling